Add arch information to the log output
Update the logging machinery to print extra data about what architecture
is being handled by which thread. This implemenation makes it
transparent for any logging user (specifically dnf in debug mode).
