+++ rpm_lockfile/__init__.py
@@ -26,7 +26,7 @@ except ImportError:
     sys.exit(127)
 import yaml
 
-from . import containers, content_origin, schema, utils
+from . import containers, content_origin, log, schema, utils
 
 CONTAINERFILE_HELP = """
 Load installed packages from base image specified in Containerfile and make
@@ -278,7 +278,7 @@ def process_arch(
     install_weak_deps: bool,
     upgrade_packages: set[str],
 ):
-    logging.info("Running solver for %s", arch)
+    log.set_thread_arch(arch)
 
     with rpmdb(arch) as root_dir:
         packages, sources, module_metadata = resolver(
@@ -403,24 +403,6 @@ def _get_containerfile_filters(context):
     return {}
 
 
-def logging_setup(debug=False):
-
-    class ExcludeErrorsFilter(logging.Filter):
-        def filter(self, record):
-            """Only lets through log messages with log level below ERROR."""
-            return record.levelno < logging.ERROR
-
-    console_stdout = logging.StreamHandler(stream=sys.stdout)
-    console_stdout.addFilter(ExcludeErrorsFilter())
-    console_stdout.setLevel(logging.DEBUG if debug else logging.INFO)
-
-    console_stderr = logging.StreamHandler(stream=sys.stderr)
-    console_stderr.setLevel(logging.ERROR)
-
-    logging.basicConfig(
-        level=logging.DEBUG if debug else logging.INFO,
-        handlers=[console_stdout, console_stderr],
-    )
 
 
 def main():
@@ -450,7 +432,7 @@ def main():
     )
     args = parser.parse_args()
 
-    logging_setup(args.debug)
+    log.setup(args.debug)
 
     config_dir = os.path.dirname(os.path.realpath(args.infile))
     with open(args.infile) as f:
